Fixes using if() with vector expressions
Which led to fixing many more checks in the internal functions and also making the tests function better as well.

Tests are now better documented.

Tests that are not applicable are removed, e.g. `test_that(".check_lonlat() stops if more than three are requested for global"` since `lonlat` is no longer used in the users' entered args, it should remain NULL if CLIMATOLOGY is selected. Same goes for setting CLIMATOLOGY if GLOBAL set in `latlon`. This isn't possible to do any longer.